pfizer-case-study-solution
Performs an initial exploratory data analysis (EDA) on a dataset.It sets up the environment by importing necessary libraries, loads the dataset, and provides a basic understanding of the data by displaying its structure, statistical summary, and checking for any missing values. This helps lay the foundation for further analysis or modeling.